feat: initial work on document proof verification
