+++ packages/rs-drive/src/drive/identity/withdrawals/documents.rs
@@ -73,7 +73,7 @@ impl Drive {
                 range_clause: None,
                 equal_clauses: where_clauses,
             },
-            offset: 0,
+            offset: None,
             limit: 100,
             order_by,
             start_at: None,
@@ -159,7 +159,7 @@ impl Drive {
                 range_clause: None,
                 equal_clauses: where_clauses,
             },
-            offset: 0,
+            offset: None,
             limit: 100,
             order_by: IndexMap::new(),
             start_at: None,
